Relocate EKS cluster code
