+++ internal/cluster/distribution/eks/ekscluster/eks.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package eks
+package ekscluster
 
 import (
 	"fmt"
